size_t
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@925 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ src/texk/web2c/luatexdir/lang/texlang.c
@@ -353,7 +353,7 @@ char *hyphenation_exception(int exceptions, char *w) {
 char *exception_strings(struct tex_language *lang) {
   char *value;
   int size = 0, current =0;
-  unsigned l =0;
+  size_t l =0;
   char *ret = NULL;
   lua_State *L = Luas[0];
   if (lang->exceptions==0)
